src/services/telegram_bot/requirements.txt
python-telegram-bot==20.0
pymongo==4.8.0
urllib3
six